+++ src/components/Inputs/_inputsComponents/InputError/helpers/getInputsErrorStyle.ts
@@ -1,12 +1,12 @@
 import { CSSProperties } from "react";
 
 import { getCssPropertyAsNumber } from "../../../../../utils";
-import { FloatingInputWidth, InputLabel, LabelPercentageWidth } from "../../../../global-types";
+import { InputLabel, Percentage } from "../../../../global-types";
 
 export const getInputsErrorStyle = (
   labelType: `${InputLabel}`,
-  labelWidth: LabelPercentageWidth,
-  floatingInputWidth: FloatingInputWidth
+  labelWidth: Percentage,
+  floatingInputWidth: Percentage
 ): CSSProperties => {
   const errorIconSize = getCssPropertyAsNumber(document.body, "--error-icon-size", 16);
   const errorIconMargin = getCssPropertyAsNumber(document.body, "--error-icon-margin", 8);