+++ src/components/Inputs/index.ts
@@ -4,6 +4,7 @@ export * from "./DateField";
 export * from "./Dropdown";
 export * from "./IconField";
 export * from "./ImageField";
+export * from "./NumberField";
 export * from "./Slider";
 export * from "./Textarea";
 export * from "./Textfield";