+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Release date 31.03.2025 - Release 2.1.3
+
+- [[#201]](https://github.com/Marcin-Migdal/m-component-library/issues/201) **[PATCH]** Apply fixes and changes needed for flaner app, useAlert data: TData | undefined, labelWidth 15-90 changed to percentage 0-100, add disableTransformContent argument to addToast payload, new helper classes w-min w-max h-min h-max
+- [[#202]](https://github.com/Marcin-Migdal/m-component-library/issues/202) **[PATCH]** Improve types in Form register function, more unified Error type for inputs and form, fix error occurring when error is a object and is passed to the input
+- [[#206]](https://github.com/Marcin-Migdal/m-component-library/issues/206) **[PATCH]** Fix dropdown input bug, when options change they did not update in the input
+- [[#207]](https://github.com/Marcin-Migdal/m-component-library/issues/207) **[PATCH]** Add NumberField input component
+
 ### Release date 25.03.2025 - Release 2.1.2
 
 - [[#186]](https://github.com/Marcin-Migdal/m-component-library/issues/186) **[PATCH]** Fix multiple bug 186, 187, 191, export FormErrorsType, Form register and registerBlur type improvement, changeDropdown input optionMenu z-index